pyproject.toml
[tool.poetry]
name = "polyga"
version = "1.0.10"
description = "Polymer Genetic Algorithm"
authors = ["Joseph Kern <jkern34@gatech.edu>", "Chiho Kim <chiho.kim@gatech.edu>"]
repository = "https://github.com/Ramprasad-Group/polyga.git"

[tool.poetry.dependencies]
python = ">=3.7.1,<3.10"
numpy = "^1.2.4"
scipy = "^1.7"
rdkit-pypi = "^2021.3.3"
pandas = "^1.1"
joblib = "^1.0.1"
SQLAlchemy = "^1.4.23"

[tool.poetry.dev-dependencies]
pydoc-markdown = ">=4.0.0,<5.0.0"
mkdocs = "^1.2.2"
pytest = "^6.2.5"
twine = "^3.4.2"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
